feat: add stdlib smt collection
This commit introduces a `collections` module for the `stdlib`. It
contains, initially, functions to support Sparse Merkle Tree
functionality. Initially, the `smt::get` is available; it will fetch the
value of a key from a Sparse Merkle tree.

It adds a `adv.smtget` that will push to the advice stack information
about a Sparse Merkle tree keyed leaf.